Fix compiler warnings (#916)
- Wsign-conversion
- Wshadow
- resolves #911

also some minor refactors

- use `std::array<GLuint,3>` typedef for `Face` in rendering, ensure correct vertex order
- add `sizeofGLVector` helper function to get size of vector of GL types
- use make_unique
- use latest manylinux2_28 wheel for readthedocs build

+++ core/common/src/tiff.cpp
@@ -238,7 +238,9 @@ TiffReader::TiffReader(const std::string &filename) {
               width, height, QImage::Format_ARGB32_Premultiplied)};
           for (int y = 0; y < qImage.height(); y++) {
             for (int x = 0; x < qImage.width(); x++) {
-              qImage.setPixel(x, y, tiffValues[y * width + x]);
+              qImage.setPixel(x, y,
+                              tiffValues[static_cast<std::size_t>(y) * width +
+                                         static_cast<std::size_t>(x)]);
             }
           }
         } else {
